requirements.txt
pyyaml
python-decouple